Refactor filter components into $lib/ui/filter
- Extract FilterSubmenu.svelte as reusable submenu component
- Move FilterDropdown.svelte and FilterSubmenu.svelte to $lib/ui/filter/
- Add svelte brand colors for hover/focus states
- Keep parent item highlighted when navigating submenu
